requirements.txt
langchain
langchain_mistralai
python-dotenv